SEACAS: Sync with main seacas repository:
commit 3493687396c8ed09ff22968c4eae62a46b1fb1fe

    IOSS: Replace vector code

commit 6e261a70c981ec76d298db7d26bc2f40f9a19a59

    APREPRO: Remove getdelim since no longer used (it is in getline.c now)

commit 1e5ac9b56e5c95f344c971f9d495549dabac889b

    NEM_SLICE: Remove sandia copyright notice since code is copyright FSF with LGPL license

commit 020e2ef3988c4815c0c9f61245e34c2d6920c161

    NEM_SLICE: Add braces to avoid compiler warning

commit c23fac8e4fd6ae932a8c3de51e6d148857fd3424

    APREPRO: Remove code with incompatible licenses; replace with BSD

commit 04dcac465b328aaea7e93d14bac4b9be6d260b53

    APREPRO: Remove obsolete COPYING file

commit 80ae0def01e54f53fe6956754eb3fc413a78c503

    Replace f2kcli arg processing routines to avoid licensing questions

commit ad06f1211698a1eabf8ec091f2d19b8fc10e3e7e

    Replace GPL getopt routines with LGPL versions

commit 5682730dea58f7f3b50578907fda0ff800ba3179

    IOSS: Remove unused variables

commit 6924f0a3c975e344ed4894107427c8cf380e33e9

    IOSS: Remove unused static function

commit 4546451dc7f3e59938f5cdaf7307cc054269d582

    GROPE: Better 1d mesh support

commit 2e93ed4ad0ed3e4dd5e870ea1f3d096c46bfb8c8

    EPU: Remove unused variables

commit dcb6f1715446c745c4bd2cfde7fa614aa513d2e7

    IOSS: Add fiels suffix option to io_info utility

commit 20a8327f20c2c6f76a6a1dc6b8c39dd128e8efd7

    GREPOS: Support for 1d models

commit 208909a10b8eff856e0a5d2e37b7df3534a6ce6c

    EPU: Remove unused variables

commit 6ef52480fb339fef13fa1edbe9ccc16976b12363

    SUPLIB: Read/write coordinates working for 1d problems

commit da2ed7fc63663e46c3c8a6340b360ed96a6f75cd

    EXOTXT/TXTEXO: Work for 1d problems

commit 3c0ac68a6d6e8cb28df67d98768fa9742774edf4

    IOSS: Add connectivty_raw output for element blocks

commit 60be6505a0cb31029b78fc3fb868f68f8a26e0ee

    IOSS: simplify hex volume calculation

commit 47680030be541239f04f0b72c7a0946018a498ae

    IOSS: Replace fortran sphgen3d with C++ Ioss version sphgen

commit e83b66ea1e39e5a8bb2f29b923b612503d0a909b

    NEMSLICE: Decide whether to open netcdf4 or netcdf3 at runtime instead of compile time

commit 922c1948d05e4bf31105cf6e8b9bd686289b735d

    NEMESIS: Decide whether to open netcdf4 or netcdf3 at runtime instead of compile time

commit bfbda2fde0a798cf2e7195c7bfe4a999d929d188

    FASTQ: If error creating output file in batch mode, do not loop forever retrying

commit 7ad60856a2529b6cf46020ca5d9c163236c0754c

    EXODUS: Add error detection for netcdf4 not being enabled in library

commit ef4e1169181ba063dc0cf1158c31bcf00c1d38b1

    EXODUS: Convert tabs to spaces

commit 4576f431e28075544bc12fbc0a51efaef41da3a0

    EPU: Do not write node/element maps it sequential

commit 3bf5a87bf67cbeed8c592ab66fdb73e1d35dac76

    EPU: Do not write node/element maps it sequential

commit 39f73aabd9add1435b3038c755af571d6a76b6a2

    EXODUS: addrwrap fixes to support large meshes

commit 8cbfe8724f8a712ce7459114fa1b37ebcb09267f

    EJOIN: Reformat help message for command-line option

commit 6da713905caec300c45b06160c6c3eff67127729

    IOSS: Avoid accessing [0]th element of empty vector

commit 6f9d5bf2bfab8bfbacafd35718180b6a96f2719d

    EXODUS: Add function prototypes

commit 1b8dfeedef2d043690be300b14af78cb483c7be7

    EXODUS: Clean up fortran-to-c interface routines

commit cb003c100d2801f4e3e2e3e0a2109913bc15a4a0

    EXODUS: Fix function name in error message

commit 538dc7b54cde256b1c970d5d74283a9eb044e21a

    EXODIFF: Fix some integer overflows for large model support

commit 12deaee0a0469f03a58d843eb9f2537cf1c77e7d

    EJOIN: Add support for transferring information records

commit 645f6af716f8c8486680ea41b46171686d566483

    IOSS: Add information record read/write support

commit f4cb4af6d7af4714aea9f85569a930abd06979e2

    MAPVAR: Initialize solution array prior to use

commit 9e2940dacd64b66047fe1256b0a72dde7a8f8098

    IOSS: Fix POINTER property retrieval

commit ad1570fc0e0ffe6cf146cb0d8ad3080e57f2b76a

    IOSS: Clean up db output formatting

commit 12d1f81fc158b4c8d4603519821c3f834aae7a25

    IOSS: Formatting changes; large model changes

commit cca40679123769ab5d9b5d43b527ab5434bb9a2e

    IOSS: Fix some indices that should be size_t instead of int

commit 815e8b84b16488f4e07a228d67265e150230ea53

    EXODIFF: Check if values need sorting prior to sorting

commit ef73e5fce4c934ca4bc0563269e19065a16bb826

    EPU: Remove check for data-set size > 2GB

commit 28fc4f53ffebb3096440decd4de0817e1fcc093c

    EXODUS: Fix function name in error message

commit ad07f99ff2f59db8e422e6e6c1b445da4033899e

    NEMESIS: Mods to use netcdf4 nonclassic if the netcdf has it; use netcdf conversions

commit 194941e8fa0a954f40cde77371c18d5fe373158f

    IOSS: Fix coordinate corruption for read/write of 1D meshes

+++ packages/seacas/applications/exodiff/exoII_read.h
@@ -78,8 +78,8 @@ public:
   
   const std::string& Title() const { return title;        }
   int     Dimension()       const { return dimension;     }
-  int     Num_Nodes()       const { return num_nodes;     }
-  int     Num_Elmts()       const { return num_elmts;     }
+  size_t  Num_Nodes()       const { return num_nodes;     }
+  size_t  Num_Elmts()       const { return num_elmts;     }
   int     Num_Node_Sets()   const { return num_node_sets; }
   int     Num_Side_Sets()   const { return num_side_sets; }
   float Data_Base_Version() const { return db_version;    }
@@ -213,10 +213,10 @@ protected:
   // GENESIS info:
   
   std::string  title;
-  int          num_nodes;
+  size_t       num_nodes;
   std::vector<std::string> coord_names;
   int          dimension;
-  int          num_elmts;
+  size_t       num_elmts;
   int          num_elmt_blocks;
   int          num_node_sets;
   int          num_side_sets;
