+++ packages/seacas/libraries/nemesis/ne_pnm.c
@@ -72,7 +72,7 @@ int ne_put_node_map(int  neid,
   int     status, varid, dimid;
   char    ftype[2];
   size_t  start[1], count[1];
-  size_t  varidx[2];
+  int64_t varidx[2];
   int  nmstat;
 
   char   errmsg[MAX_ERR_LENGTH];
